requirements.txt
torch
numpy
Pillow